+++ packages/connectors/src/walletConnect.ts
@@ -1,34 +1,56 @@
-import type { ProviderRpcError } from '@wagmi/core'
 import {
+  Chain,
+  ProviderRpcError,
   SwitchChainError,
   UserRejectedRequestError,
   getClient,
   normalizeChainId,
 } from '@wagmi/core'
-import type { Chain } from '@wagmi/core/chains'
 import type WalletConnectProvider from '@walletconnect/ethereum-provider'
+import type {
+  UniversalProviderOpts,
+  UniversalProvider as UniversalProvider_,
+} from '@walletconnect/universal-provider'
 import { providers } from 'ethers'
 import { getAddress, hexValue } from 'ethers/lib/utils.js'
 
 import { Connector } from './base'
 
-/**
- * Wallets that support chain switching through WalletConnect
- * - imToken (token.im)
- * - MetaMask (metamask.io)
- * - Rainbow (rainbow.me)
- * - Trust Wallet (trustwallet.com)
- */
-const switchChainAllowedRegex = /(imtoken|metamask|rainbow|trust wallet)/i
+// Shared config for WalletConnect v2
+const defaultV2Config = {
+  namespace: 'eip155',
+  methods: [
+    'eth_sendTransaction',
+    'eth_sign',
+    'eth_signTransaction',
+    'eth_signTypedData',
+    'personal_sign',
+  ],
+  events: ['accountsChanged', 'chainChanged'],
+}
 
-type WalletConnectOptions = ConstructorParameters<
-  typeof WalletConnectProvider
->[0]
+type UniversalProvider = InstanceType<typeof UniversalProvider_>
+
+type WalletConnectOptions = {
+  /** When `true`, uses default WalletConnect QR Code modal */
+  qrcode?: boolean
+} & (
+  | ({
+      version?: '1'
+    } & ConstructorParameters<typeof WalletConnectProvider>[0])
+  | ({
+      /**
+       * Project ID associated with [WalletConnect account](https://cloud.walletconnect.com)
+       */
+      projectId: NonNullable<UniversalProviderOpts['projectId']>
+      version: '2'
+    } & Omit<UniversalProviderOpts, 'projectId'>)
+)
 
 type WalletConnectSigner = providers.JsonRpcSigner
 
 export class WalletConnectConnector extends Connector<
-  WalletConnectProvider,
+  WalletConnectProvider | UniversalProvider,
   WalletConnectOptions,
   WalletConnectSigner
 > {
@@ -36,10 +58,15 @@ export class WalletConnectConnector extends Connector<
   readonly name = 'WalletConnect'
   readonly ready = true
 
-  #provider?: WalletConnectProvider
+  #provider?: WalletConnectProvider | UniversalProvider
 
   constructor(config: { chains?: Chain[]; options: WalletConnectOptions }) {
     super(config)
+  }
+
+  get version() {
+    if ('version' in this.options) return this.options.version
+    return '1'
   }
 
   async connect({ chainId }: { chainId?: number } = {}) {
@@ -59,19 +86,108 @@ export class WalletConnectConnector extends Connector<
       provider.on('chainChanged', this.onChainChanged)
       provider.on('disconnect', this.onDisconnect)
 
+      if (this.version === '2') {
+        provider.on('session_delete', this.onDisconnect)
+        provider.on('display_uri', this.onDisplayUri)
+
+        const isChainsAuthorized = await this.#isChainsAuthorized()
+
+        // If there is an active session, and the chains are not authorized,
+        // disconnect the session.
+        if ((provider as UniversalProvider).session && !isChainsAuthorized)
+          await provider.disconnect()
+
+        // If there is not an active session, or the chains are not authorized,
+        // attempt to connect.
+        if (
+          !(provider as UniversalProvider).session ||
+          ((provider as UniversalProvider).session && !isChainsAuthorized)
+        ) {
+          await Promise.race([
+            provider.connect({
+              namespaces: {
+                [defaultV2Config.namespace]: {
+                  methods: defaultV2Config.methods,
+                  events: defaultV2Config.events,
+                  chains: this.chains.map(
+                    (chain) => `${defaultV2Config.namespace}:${chain.id}`,
+                  ),
+                  rpcMap: this.chains.reduce(
+                    (rpc, chain) => ({
+                      ...rpc,
+                      [chain.id]: chain.rpcUrls.default.http[0],
+                    }),
+                    {},
+                  ),
+                },
+              },
+            }),
+            ...(this.options.qrcode
+              ? // When using WalletConnect QR Code Modal, open modal and listen for close callback.
+                // If modal is closed, reject promise so `catch` block for `connect` is called.
+                [
+                  new Promise((_res, reject) =>
+                    provider.on('display_uri', async (uri: string) =>
+                      (
+                        await import('@walletconnect/qrcode-modal')
+                      ).default.open(uri, () =>
+                        reject(new Error('user rejected')),
+                      ),
+                    ),
+                  ),
+                ]
+              : []),
+          ])
+
+          // If execution reaches here, connection was successful and we can close modal.
+          if (this.options.qrcode)
+            (await import('@walletconnect/qrcode-modal')).default.close()
+        }
+      }
+
       // Defer message to the next tick to ensure wallet connect data (provided by `.enable()`) is available
       setTimeout(() => this.emit('message', { type: 'connecting' }), 0)
 
-      const accounts = await provider.enable()
+      const accounts = (await Promise.race([
+        provider.enable(),
+        // When using WalletConnect v1 QR Code Modal, handle user rejection request from wallet
+        ...(this.version === '1' && this.options.qrcode
+          ? [
+              new Promise((_res, reject) =>
+                (provider as WalletConnectProvider).connector.on(
+                  'disconnect',
+                  () => reject(new Error('user rejected')),
+                ),
+              ),
+            ]
+          : []),
+      ])) as string[]
       const account = getAddress(accounts[0] as string)
       const id = await this.getChainId()
       const unsupported = this.isChainUnsupported(id)
 
-      // Not all WalletConnect options support programmatic chain switching
-      // Only enable for wallet options that do
-      const walletName = provider.connector?.peerMeta?.name ?? ''
-      if (switchChainAllowedRegex.test(walletName))
-        this.switchChain = this.#switchChain
+      // Not all WalletConnect v1 options support programmatic chain switching.
+      // Only enable for wallet options that do.
+      if (this.version === '1') {
+        const walletName =
+          (provider as WalletConnectProvider).connector?.peerMeta?.name ?? ''
+
+        /**
+         * Wallets that support chain switching through WalletConnect
+         * - imToken (token.im)
+         * - MetaMask (metamask.io)
+         * - Omni (omni.app)
+         * - Rainbow (rainbow.me)
+         * - Trust Wallet (trustwallet.com)
+         */
+        const switchChainAllowedRegex =
+          /(imtoken|metamask|omni|rainbow|trust wallet)/i
+        if (switchChainAllowedRegex.test(walletName))
+          this.switchChain = this.#switchChain
+      }
+      // In v2, chain switching is allowed programatically
+      // as the user approves the chains when approving the pairing
+      else this.switchChain = this.#switchChain
 
       return {
         account,
@@ -81,35 +197,69 @@ export class WalletConnectConnector extends Connector<
         ),
       }
     } catch (error) {
-      if (/user closed modal/i.test((error as ProviderRpcError).message))
+      if (this.version === '2' && this.options.qrcode)
+        (await import('@walletconnect/qrcode-modal')).default.close()
+      // WalletConnect v1: "user closed modal"
+      // WalletConnect v2: "user rejected"
+      if (
+        /user closed modal|user rejected/i.test(
+          (error as ProviderRpcError)?.message,
+        )
+      ) {
         throw new UserRejectedRequestError(error)
+      }
       throw error
     }
   }
 
   async disconnect() {
     const provider = await this.getProvider()
-    await provider.disconnect()
+    try {
+      await provider.disconnect()
+    } catch (error) {
+      // If the session does not exist, we don't want to throw.
+      if (!/No matching key/i.test((error as Error).message)) throw error
+    }
 
     provider.removeListener('accountsChanged', this.onAccountsChanged)
     provider.removeListener('chainChanged', this.onChainChanged)
     provider.removeListener('disconnect', this.onDisconnect)
 
-    typeof localStorage !== 'undefined' &&
+    if (this.version === '1' && typeof localStorage !== 'undefined')
+      // Remove local storage session so user can connect again
       localStorage.removeItem('walletconnect')
+    else {
+      provider.removeListener('session_delete', this.onDisconnect)
+      provider.removeListener('display_uri', this.onDisplayUri)
+    }
   }
 
   async getAccount() {
     const provider = await this.getProvider()
-    const accounts = provider.accounts
+    let accounts
+    if (this.version === '1')
+      accounts = (provider as WalletConnectProvider).accounts
+    else
+      accounts = (await provider.request({
+        method: 'eth_accounts',
+      })) as string[]
+
     // return checksum address
     return getAddress(accounts[0] as string)
   }
 
   async getChainId() {
     const provider = await this.getProvider()
-    const chainId = normalizeChainId(provider.chainId)
-    return chainId
+    if (this.version === '1')
+      return normalizeChainId((provider as WalletConnectProvider).chainId)
+
+    // WalletConnect v2 does not internally manage chainIds anymore, so
+    // we need to retrieve it from the client, or request it from the provider
+    // if none exists.
+    return (
+      getClient().data?.chain?.id ??
+      normalizeChainId(await provider.request({ method: 'eth_chainId' }))
+    )
   }
 
   async getProvider({
@@ -118,24 +268,39 @@ export class WalletConnectConnector extends Connector<
   }: { chainId?: number; create?: boolean } = {}) {
     // Force create new provider
     if (!this.#provider || chainId || create) {
-      const rpc = !this.options?.infuraId
-        ? this.chains.reduce(
-            (rpc, chain) => ({
-              ...rpc,
-              [chain.id]: chain.rpcUrls.default.http[0],
-            }),
-            {},
+      if (this.options.version === '2') {
+        const WalletConnectProvider = (
+          await import('@walletconnect/universal-provider')
+        ).default
+        this.#provider = await WalletConnectProvider.init(
+          this.options as UniversalProviderOpts,
+        )
+        if (chainId)
+          this.#provider.setDefaultChain(
+            `${defaultV2Config.namespace}:${chainId}`,
           )
-        : {}
+        return this.#provider
+      } else {
+        const rpc = !this.options?.infuraId
+          ? this.chains.reduce(
+              (rpc, chain) => ({
+                ...rpc,
+                [chain.id]: chain.rpcUrls.default.http[0],
+              }),
+              {},
+            )
+          : {}
 
-      const WalletConnectProvider = (
-        await import('@walletconnect/ethereum-provider')
-      ).default
-      this.#provider = new WalletConnectProvider({
-        ...this.options,
-        chainId,
-        rpc: { ...rpc, ...this.options?.rpc },
-      })
+        const WalletConnectProvider = (
+          await import('@walletconnect/ethereum-provider')
+        ).default
+        this.#provider = new WalletConnectProvider({
+          ...this.options,
+          chainId,
+          rpc: { ...rpc, ...this.options?.rpc },
+        })
+        return this.#provider
+      }
     }
 
     return this.#provider
@@ -146,19 +311,55 @@ export class WalletConnectConnector extends Connector<
       this.getProvider({ chainId }),
       this.getAccount(),
     ])
-    return new providers.Web3Provider(
-      provider as providers.ExternalProvider,
-      chainId,
-    ).getSigner(account)
+
+    let provider_ = provider as providers.ExternalProvider
+    if (this.version === '2') {
+      const chainId_ = await this.getChainId()
+      provider_ = {
+        ...provider,
+        async request(args) {
+          return await provider.request(
+            args,
+            `${defaultV2Config.namespace}:${chainId ?? chainId_}`,
+          )
+        },
+      } as providers.ExternalProvider
+    }
+
+    return new providers.Web3Provider(provider_, chainId).getSigner(account)
   }
 
   async isAuthorized() {
     try {
-      const account = await this.getAccount()
-      return !!account
+      const [account, isChainsAuthorized] = await Promise.all([
+        this.getAccount(),
+        this.#isChainsAuthorized(),
+      ])
+      return !!account && isChainsAuthorized
     } catch {
       return false
     }
+  }
+
+  /**
+   * @description Checks if the connector is authorized to access the requested chains.
+   *
+   * There could be a case where requested chains are out of sync with
+   * the users' approved chains (e.g. the dapp could have added support
+   * for an additional chain), hence the user will be unauthorized.
+   */
+  async #isChainsAuthorized() {
+    const provider = await this.getProvider()
+
+    if (this.version === '1') return true
+
+    const providerChains =
+      (provider as UniversalProvider).namespaces?.[defaultV2Config.namespace]
+        ?.chains || []
+    const authorizedChainIds = providerChains.map(
+      (chain) => parseInt(chain.split(':')[1] || '') as Chain['id'],
+    )
+    return !this.chains.some(({ id }) => !authorizedChainIds.includes(id))
   }
 
   async #switchChain(chainId: number) {
@@ -181,12 +382,18 @@ export class WalletConnectConnector extends Connector<
           }),
         ),
       ])
+      if (this.version === '2') {
+        ;(provider as UniversalProvider).setDefaultChain(
+          `${defaultV2Config.namespace}:${chainId}`,
+        )
+        this.onChainChanged(chainId)
+      }
       return (
         this.chains.find((x) => x.id === chainId) ?? {
           id: chainId,
           name: `Chain ${id}`,
           network: `${id}`,
-          nativeCurrency: { name: 'Ether', decimals: 18, symbol: 'ETH' },
+          nativeCurrency: { decimals: 18, name: 'Ether', symbol: 'ETH' },
           rpcUrls: { default: { http: [''] } },
         }
       )
@@ -212,5 +419,9 @@ export class WalletConnectConnector extends Connector<
 
   protected onDisconnect = () => {
     this.emit('disconnect')
+  }
+
+  protected onDisplayUri = (uri: string) => {
+    this.emit('message', { type: 'display_uri', data: uri })
   }
 }